split BE & FE
