--- .env
@@ -1,2 +1,0 @@
-CLIENT_PORT=8000
-SERVER_PORT=9000
