+++ .gitignore
@@ -1,22 +1,9 @@
-# dependencies
 */node_modules
-/.pnp
-.pnp.js
-
-# testing
-/client/coverage
-
-# production
+.vscode/*
+.idea/*
 /client/build
-
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
+/server/.serverless/*
 
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-.idea/*
