+++ client/src/config/endpoints.ts
@@ -1,2 +1,2 @@
 const baseUrl = 'http://localhost:5000';
-export const landingPageUrl = baseUrl + '/';
+export const helloWorldApi = baseUrl + '/helloWorld';
