+++ server/app.sls.js
@@ -1,0 +1,4 @@
+const sls = require('serverless-http');
+const app = require('./app');
+
+module.exports.server = sls(app);
