+++ server/config/env.json
@@ -1,0 +1,4 @@
+{
+  "CLIENT_PORT":"3000",
+  "SERVER_PORT": "5000"
+}
