+++ server/secrets.json
@@ -1,0 +1,3 @@
+{
+  "NODE_ENV": "production"
+}
